Move the C++ classes and definitions of inline functions from the Mapper.cpp file to the Mapper.h file
